Add new APIs, use best possible scheduler
All new APIs from 24 and 26 are supported except:
- ClipData and related
- JobWorkItem and related

JobSchedulerCompat is now also using the best possible scheduler
for each job, and not the same scheduler for all jobs. Given all
new APIs this made sense, as users can easily use APIs which are
unavailable in certain JobScheduler implementations
(eg. flexMillis in API 21) and using a lesser scheduler for all
jobs would be a shame.