Fix accessing JobStore after a job has been deleted
Between API 21 and 26, if a job is canceled while running,
JobSchedulerJobService would fetch the job from the store to grab its
transient extras. However, the job had already been deleted.

Instead of relying in JobStore throughout the workflow, which is prone
to race conditions, transient extras are now fetched upon starting,
which is the only time we can guarantee the job exists.